Add 알람 AI 진단 + 지문(fingerprint) 그룹핑 수정
알람 하나를 열고 계정을 지정하면 그 계정에서 읽기 전용 조회를 돌려
짧은 진단을 돌려주는 기능을 추가한다.

지문 수정 (api/normalize_handler.py):
메시지를 그대로 해시해서 "CPU 92%" 와 "CPU 87%" 가 다른 지문이 되었다.
같은 알람이 매번 다른 종류로 취급되어 억제 규칙도 발생 횟수 집계도
성립하지 않았다. 변하는 값(숫자, IP, UUID, ARN, 리소스 ID, 타임스탬프,
긴 16진수)을 자리표시자로 바꾼 '틀' 을 해시한다. meta 에 AlarmName /
alertname 이 있으면 메시지 대신 그 이름으로 묶는다.
샘플 400건 기준 지문 400종 -> 192종.

진단 (app/agent_core.py, app/views/alarm.py):
- 대화가 아니라 요청 한 번. 기록을 남기지 않으므로 고객사 A 의 조회
  결과가 고객사 B 의 진단에 섞이지 않는다.
- 조회 도구가 받는 인자는 command 하나뿐. 계정과 리전은 담당자가 화면에서
  고른 값이 flask.g 로 전달된다. 계정을 도구 인자로 두면 모델이 다른
  고객사의 계정 번호를 지어내 조회할 수 있다.
- 실행 계층은 콘솔과 같은 코드. 사람이 친 명령이든 모델이 만든 명령이든
  같은 허용 목록을 통과하므로 신뢰 경계가 바뀌지 않는다.
- 거부는 예외가 아니라 문자열로 돌려준다. 예외로 던지면 루프가 죽고
  모델은 이유를 알 수 없다. 반복 시도는 max_iterations 로 끊는다.
- 모델에게 넘기는 출력 상한은 12,000자(콘솔은 20만자). 잘리면 --query 로
  좁혀 다시 조회하라는 안내를 함께 준다.
- 발생 이력(stats.fingerprint_history)을 프롬프트에 함께 넣는다.
  처음 발생인지 사흘째 반복인지에 따라 봐야 할 곳이 달라진다.
- 감사 기록은 진단 한 건당 하나. meta.actor=agent 로 사람이 친 명령과
  구분한다.

그 외:
- alarm.html: meta 가 200자를 넘으면 접어둔다(감사 기록이 목록을 밀어냄)
- report_pptx.py: 표 좌표를 int 로 정리. 호출부의 나눗셈이 float 를
  만들어 python-pptx 가 거부했다(/report/download.pptx 500 에러)

+++ myapp/app/report_pptx.py
@@ -132,12 +132,17 @@ def _delta_text(change):
 def _table(slide, left, top, width, rows, headers, col_widths=None, row_h=Inches(0.32)):
     """간단한 표. python-pptx 기본 표 스타일은 요란해서 직접 칠한다."""
     n_rows, n_cols = len(rows) + 1, len(headers)
-    shape = slide.shapes.add_table(n_rows, n_cols, left, top, width, row_h * n_rows)
+    # 길이(EMU)는 정수여야 한다. 호출부에서 나눗셈으로 폭을 계산하면
+    # 파이썬 / 는 float 를 돌려주고, python-pptx 는 그걸 그대로 거부한다
+    # (TypeError: value must be an integral type). 여기서 한 번 정리한다.
+    shape = slide.shapes.add_table(
+        n_rows, n_cols, int(left), int(top), int(width), int(row_h * n_rows)
+    )
     table = shape.table
 
     if col_widths:
         for i, cw in enumerate(col_widths):
-            table.columns[i].width = cw
+            table.columns[i].width = int(cw)
 
     for r in range(n_rows):
         table.rows[r].height = row_h
